Convert project to Vue 3 with Vuetify
