+++ index.html
@@ -4,12 +4,9 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Procedural Planet</title>
-  <style>
-    body { margin: 0; overflow: hidden; }
-    canvas { display: block; }
-  </style>
 </head>
 <body>
+  <div id="app"></div>
   <script type="module" src="/src/main.js"></script>
 </body>
 </html>
