+++ package.json
@@ -11,10 +11,14 @@
   "dependencies": {
     "three": "^0.158.0",
     "express": "^4.19.2",
-    "cors": "^2.8.5"
+    "cors": "^2.8.5",
+    "vue": "^3.4.0",
+    "vuetify": "^3.5.0"
   },
   "devDependencies": {
-    "vite": "^5.0.0"
+    "vite": "^5.0.0",
+    "@vitejs/plugin-vue": "^4.5.0",
+    "vite-plugin-vuetify": "^1.0.0"
 
   },
   "author": "",
