+++ src/App.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-app>
+    <PlanetViewer />
+  </v-app>
+</template>
+
+<script setup>
+import PlanetViewer from './components/PlanetViewer.vue';
+</script>
