+++ src/noise.js
@@ -1,8 +1,6 @@
 // Simple deterministic noise function using a seed.
 // Based on sine pseudorandom generation.
-function noise(x, seed) {
+export default function noise(x, seed) {
   const sin = Math.sin(x * seed) * 10000;
   return sin - Math.floor(sin);
 }
-
-module.exports = noise;
