+++ vite.config.js
@@ -1,0 +1,7 @@
+import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
+import vuetify from 'vite-plugin-vuetify';
+
+export default defineConfig({
+  plugins: [vue(), vuetify()]
+});
